refactor: separate engine api
