+++ Pandora API/IOManagers/IMetaDataExporter.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Pandora.API.Patch.IOManagers;
+
+public interface IMetaDataExporter<T> : IDataExporter<T>
+{
+	public void LoadMetaData();
+	public void SaveMetaData(IEnumerable<T> collection);
+}
+
