+++ Pandora API/Patch.Engine.Config/IEngineConfiguration.cs
@@ -1,0 +1,11 @@
+﻿namespace Pandora.API.Patch.Engine.Config;
+
+public interface IEngineConfiguration
+{
+    string Name { get; }
+
+    string Description { get; }
+
+    public IPatcher Patcher { get; }
+
+}
