+++ Pandora Behaviour Engine/Data/IModInfoProvider.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Pandora.Core;
-
+using Pandora.API.Patch;
 namespace Pandora.MVVM.Data;
 
 public interface IModInfoProvider
