+++ Pandora Behaviour Engine/Models/FNISModInfo.cs
@@ -1,3 +1,4 @@
+using Pandora.API.Patch;
 using System;
 using System.Collections.Generic;
 using System.IO;
