--- Pandora Behaviour Engine/Models/IModInfo.cs
@@ -1,42 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Diagnostics.Contracts;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Pandora.Core
-{
-    /// <summary>
-    /// Strongly recommended to implement GetHashCode() in addition to IEquatable.
-    /// </summary>
-	public interface IModInfo : IEquatable<IModInfo>
-    {
-        public enum ModFormat
-        {
-            FNIS, 
-            Nemesis, 
-            Pandora
-        }
-
-        public string Name { get; }
-
-        public string Author { get; }
-
-        public string URL { get; }
-
-        public string Code { get; }
-
-        public Version Version { get; }
-
-        public DirectoryInfo Folder { get; }
-
-        public ModFormat Format { get; }
-
-		public bool Active { get; set;  }
-
-        public uint Priority { get; set;  }
-	}
-}
