+++ Pandora Behaviour Engine/Models/Patch/IOManagers/Skyrim/PackFileExporter.cs
@@ -1,5 +1,6 @@
 ﻿using HKX2E;
 using NLog;
+using Pandora.API.Patch.IOManagers;
 using Pandora.Core;
 using Pandora.Patch.Patchers.Skyrim.Hkx;
 using System;
