+++ Pandora Behaviour Engine/Models/Patch/Patchers/Skyrim/Hkx/PackFileChangeSet.cs
@@ -1,4 +1,5 @@
 ﻿using NLog;
+using Pandora.API.Patch;
 using Pandora.Core;
 using System;
 using System.Collections.Generic;
