+++ Pandora Behaviour Engine/Models/Patch/Patchers/Skyrim/Hkx/PatchNodeCreator.cs
@@ -1,4 +1,5 @@
 ﻿using HKX2E;
+using Pandora.API.Patch;
 using Pandora.Core;
 using System;
 using System.Collections.Generic;
