+++ Pandora Behaviour Engine/ViewModels/EngineConfigurationViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using NLog.Filters;
+using Pandora.API.Patch.Engine.Config;
 using Pandora.Command;
 using Pandora.Core;
 using Pandora.Core.Engine.Configs;
@@ -12,12 +13,7 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace Pandora.ViewModels;
-public interface IEngineConfigurationViewModel : INotifyPropertyChanged
-{
-	public string Name { get; }
-	public RelayCommand? SetCommand { get; }
-	public ObservableCollection<IEngineConfigurationViewModel> NestedViewModels { get; }
-}
+
 public class EngineConfigurationViewModel<T> : IEngineConfigurationFactory,IEngineConfigurationViewModel where T : class, IEngineConfiguration, new()
 {
     public event PropertyChangedEventHandler? PropertyChanged;
@@ -36,25 +32,6 @@ public class EngineConfigurationViewModel<T> : IEngineConfigurationFactory,IEngi
 	{
 		Name = name;
 		SetCommand = setCommand;
-	}
-}
-
-public class EngineConfigurationViewModelContainer : IEngineConfigurationViewModel
-{
-	public event PropertyChangedEventHandler? PropertyChanged;
-	private void RaisePropertyChanged([CallerMemberName] string? propertyName = null)
-	{
-		PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-	}
-	public string Name { get; private set; }
-
-	public RelayCommand? SetCommand { get; } = null;
-
-	public ObservableCollection<IEngineConfigurationViewModel> NestedViewModels { get; private set; } = new ObservableCollection<IEngineConfigurationViewModel>();
-	public EngineConfigurationViewModelContainer(string name, params IEngineConfigurationViewModel[] viewModels)
-	{
-		Name = name;
-		foreach (var viewModel in viewModels) { NestedViewModels.Add(viewModel); }
 	}
 }
