+++ Pandora+.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XML Cake", "XML.Cake.NET\XM
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HKX2E", "HKX2 Enhanced Library\HKX2\HKX2E.csproj", "{39BEC9FE-DE80-4D85-B6C4-2BAD740624E4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pandora API", "Pandora API\Pandora API.csproj", "{9A0989E3-AE1A-4DE3-992F-B3FF96DAECAC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{39BEC9FE-DE80-4D85-B6C4-2BAD740624E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{39BEC9FE-DE80-4D85-B6C4-2BAD740624E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{39BEC9FE-DE80-4D85-B6C4-2BAD740624E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9A0989E3-AE1A-4DE3-992F-B3FF96DAECAC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9A0989E3-AE1A-4DE3-992F-B3FF96DAECAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9A0989E3-AE1A-4DE3-992F-B3FF96DAECAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9A0989E3-AE1A-4DE3-992F-B3FF96DAECAC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
